Use IdentityManager's preferred role assignment mechanism.
See:
https://github.com/IdentityManager/IdentityManager.AspNetIdentity/issues/21.

In a nutshell, the IdentityServer folks maintain that Asp.Net Identity's
establishment of the 'Role' as a sprecial attribute is goofy relic of
times past, and that roles should be treated as any other normal claim.
With this change role updates will result in additions to th
`AspNetUserClaims` table, and not the `AspNetUserRoles` table. The
`Configuration` now behaves this way when assigning roles to default
users.